renamed fix_delay.py and deleted old search_map.py
